% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/listdemo.R
\name{listdemo}
\alias{listdemo}
\title{List the demo and select one to run}
\usage{
listdemo()
}
\value{
The app will start in a web browser
}
\description{
This function shows a list of shiny app demos for user to choose one to run.
}
\examples{
\dontrun{
listdemo()
}
}

