% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listdemos.R
\name{listdemos}
\alias{listdemos}
\title{List the demos and select one to run}
\usage{
listdemos()
}
\value{
The app will start in a web browser
}
\description{
This function shows a list of shiny app demos for user to choose one to run.
}
\examples{
\dontrun{
listdemos()
}
}

