% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rundemo.R
\name{rundemo}
\alias{rundemo}
\title{Run a demo}
\usage{
rundemo(demoname)
}
\arguments{
\item{demoname}{The name of the demo}
}
\value{
The app will start in a web browser
}
\description{
This function run a shiny app demo in this package
}
\examples{
\dontrun{
rundemo("ANOVAResidual")
}
}

