% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lstatdemo-package.R
\docType{package}
\name{lstatdemo}
\alias{lstatdemo}
\alias{lstatdemo-package}
\title{lstatdemo: Demonstrations, usually shiny apps, for learning statistics}
\description{
lstatdemo: Demonstrations, usually shiny apps, for learning statistics
}

